+++ src/classnames.ts
@@ -1,0 +1,2 @@
+import * as classNames from "https://deno.land/x/classnames@0.1.1/index.ts";
+export const cn = classNames.default;